[Feat] 필터링기능 커스텀훅 분리
