+++ src/Pages/MainPage.jsx
@@ -1,27 +1,20 @@
-import React, { useMemo, useEffect, useState } from 'react';
+import React, { useMemo, useState } from 'react';
 import styled from 'styled-components';
 import TopBar from '../Components/Common/TopBar';
 import TotalCount from '../Components/Article/TotalCount';
 import FeedContents from '../Components/Article/FeedContents';
 import BottomNav from '../Components/Common/BottomNav';
 import GetShowAPI from '../API/GetShowAPI';
-import {
-  Show,
-  IsLoginState,
-  CategoryInterestTagCount,
-  CategoryAreaTagCount,
-  InterestTags,
-  AreaTags,
-} from '../Atom/atom';
+import { Show, IsLoginState, CategoryInterestTagCount, CategoryAreaTagCount } from '../Atom/atom';
 import { useRecoilState, useRecoilValue } from 'recoil';
+import useDataFiltering from '../Hook/useDataFiltering';
 
 const MainPage = () => {
   const [getShow, setShow] = useRecoilState(Show);
   const [isLoginState, setIsLoginState] = useRecoilState(IsLoginState);
   const interestTagCount = useRecoilValue(CategoryInterestTagCount);
   const areaTagCount = useRecoilValue(CategoryAreaTagCount);
-  const interastTags = useRecoilValue(InterestTags);
-  const areaTags = useRecoilValue(AreaTags);
+
   const [showData, setShowData] = useState([]);
 
   console.log(`tags count : ${interestTagCount} / ${areaTagCount}`);
@@ -41,29 +34,7 @@ const MainPage = () => {
     }
   }, [getShow, showData]);
 
-  useEffect(() => {
-    let resultData = [...getShow];
-    if (interestTagCount || areaTagCount) {
-      let trueList = [];
-      if (interestTagCount) {
-        trueList = interastTags.filter(el => el[1] === true).map(tag => tag[0]);
-        resultData = [
-          ...getShow.filter(data => {
-            return trueList.some(tag => tag === data.CODENAME);
-          }),
-        ];
-      }
-      if (areaTagCount) {
-        trueList = areaTags.filter(el => el[1] === true).map(tag => tag[0]);
-        resultData = [
-          ...getShow.filter(data => {
-            return trueList.some(tag => tag === data.GUNAME);
-          }),
-        ];
-      }
-    }
-    setShowData(resultData);
-  }, []);
+  useDataFiltering(getShow, setShowData, interestTagCount, areaTagCount);
 
   console.log('메인피드 렌더링...', showData);
